# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: VisionLoop API
  version: 0.2.0
  summary: Phase 1 + Phase 2 surface — auth, images, jobs (with SSE), annotations.
  description: |
    Source of truth: `docs/spec/visionloop-spec.md`. Sections cited inline.

    Auth: Better Auth session cookie (`better-auth.session_token`) or
    `Authorization: Bearer <token>` (spec §9). All `/v1/*` endpoints require
    a session and a primary org membership. RBAC (owner/admin/annotator/
    viewer) is enforced per-handler — viewers are read-only.

    Multi-tenancy: every query scopes by `org_id` resolved from the caller's
    primary membership. Cross-tenant reads return 404, never 403.

    Errors share one envelope: `{ "error": { "code", "message", "request_id" } }`.
    Some errors add fields (e.g. `current_version` on 409, `retry_after_ms`
    on 503 from POST /v1/jobs).

    Internal worker callback (`/internal/jobs/callback`) is intentionally
    outside `/v1` and is guarded by a shared bearer **plus** a per-job
    HMAC over the job id (ADR-0004).
  contact:
    name: VisionLoop
    url: https://github.com/visionloop
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: http://localhost:8080
    description: Local docker-compose
  - url: https://api.visionloop.example.com
    description: Production (placeholder)
...
security:
  - sessionCookie: []
  - bearerAuth: []
paths:
...
  /v1/images:
    get:
      operationId: listImages
      tags: [images]
      summary: List images for the caller's org
      parameters:
        - in: query
          name: limit
          schema: { type: integer, format: int32, default: 50, minimum: 1, maximum: 500 }
        - in: query
          name: offset
          schema: { type: integer, format: int32, default: 0, minimum: 0 }
        - in: query
          name: status
          schema: { $ref: '#/components/schemas/ImageStatus' }
      responses:
        '200':
          description: Page of images.
          content:
            application/json:
              schema:
                type: object
                required: [items, total, limit, offset]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Image' }
                  total:  { type: integer, format: int64 }
                  limit:  { type: integer, format: int32 }
                  offset: { type: integer, format: int32 }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: better-auth.session_token
      description: |
        Better Auth session cookie. Server-side HMAC-verifies the cookie
        and resolves the `users.id` (ADR-0003).
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: better-auth-session
      description: |
        Same Better Auth session token, sent via `Authorization: Bearer`
        for non-browser clients.
    workerCallback:
      type: http
      scheme: bearer
      description: |
        Worker callback shared bearer (`JOB_CALLBACK_TOKEN`). The handler
        also requires `X-Job-ID` and `X-Job-Signature` (HMAC-SHA256 of the
        job id with `JOB_CALLBACK_HMAC_SECRET`). See ADR-0004.
...
  responses:
    Unauthorized:
      description: Missing or invalid session.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Caller's role is not allowed for this operation.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    ImageStatus:
      type: string
      enum: [uploading, ready, errored, archived]
...
    Image:
      type: object
      required: [id, org_id, status, storage_key, content_type, byte_size]
      properties:
        id:           { type: string, format: uuid }
        org_id:       { type: string, format: uuid }
        status:       { $ref: '#/components/schemas/ImageStatus' }
        storage_key:  { type: string }
        content_type: { type: string }
        byte_size:    { type: integer, format: int64 }
        width:        { type: integer, format: int32 }
        height:       { type: integer, format: int32 }
...
    ErrorBody:
      type: object
      required: [code, message, request_id]
      properties:
        code:       { type: string, examples: [unauthorized, forbidden, not_found, invalid_input, conflict, too_large, unsupported_media_type, JOB_PUBLISH_FAILED, if_match_required, if_match_invalid] }
        message:    { type: string }
        request_id: { type: string, description: Echoed `X-Request-ID`. }
    Error:
      type: object
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/ErrorBody' }